+++ tests/phpunit/Util.php
@@ -9,10 +9,13 @@
 class Util {
 
 	/**
+	 * Create a new page
+	 *
 	 * @since 0.1.5
 	 * @param string|Title $title
 	 * @param string $content
 	 * @param int $namespace Default is NS_MAIN.
+	 * @return WikiPage
 	 */
 	public static function addPage(
 		$title,
@@ -27,6 +30,28 @@ class Util {
 			ContentHandler::makeContent(
 				$content,
 				$title,
+				CONTENT_MODEL_WIKITEXT
+			),
+			''
+		);
+		return $page;
+	}
+
+	/**
+	 * Edit an existing page
+	 *
+	 * @since 0.1.5
+	 * @param WikiPage $page
+	 * @param string $content
+	 */
+	public static function editPage(
+		$page,
+		$content
+	) {
+		$page->doEditContent(
+			ContentHandler::makeContent(
+				$content,
+				$page->getTitle(),
 				CONTENT_MODEL_WIKITEXT
 			),
 			''