+++ CHANGES.md
@@ -28,6 +28,7 @@ the match the new release version.
 ### 0.1.5-SNAPSHOT
 202Y-MM-DD
 
+* [T260891](https://phabricator.wikimedia.org/T260891) Allow historical revisions to be accessed in Segmenter cache.
 * [T260875](https://phabricator.wikimedia.org/T260875) Enable using default parameters for segmenting.
 * [T255001](https://phabricator.wikimedia.org/T255001) Schedulable cleanup job of orphaned utterance files in file backend.
 * [T257571](https://phabricator.wikimedia.org/T257571) Extension is owner of default voice per language logic.